Type cast to correct data type for comparision in NANTLV_ReadTlv
pOutTlv->length is of type u16 while the inBufferSize is int.
The comparision between these different types is incorrect.
Hence, Type cast to correct data type for comparision in
NANTLV_ReadTlv.

CRs-fixed: 2884038

+++ qcwcn/wifi_hal/nan.cpp
@@ -1425,7 +1425,7 @@ u16 NANTLV_ReadTlv(u8 *pInTlv, pNanTlv pOutTlv, int inBufferSize)
     pOutTlv->length |= *pInTlv++ << 8;
     readLen += 2;
 
-    if(pOutTlv->length > inBufferSize - NAN_TLV_HEADER_SIZE) {
+    if(pOutTlv->length > (u16)(inBufferSize - NAN_TLV_HEADER_SIZE)) {
         ALOGE("Insufficient length to process TLV header, inBufferSize = %d",
               inBufferSize);
         return readLen;
